NeighborlyAPI/requirements.txt
azure-functions==1.5.0
pymongo==3.11.2
